feat(frontend): Refactor app structure by separating App component and index entry point
