+++ frontend/src/App.tsx
@@ -1,5 +1,3 @@
-import { StrictMode } from 'react'
-import ReactDOM from 'react-dom/client'
 import { RouterProvider, createRouter } from '@tanstack/react-router'
 import { QueryClient, QueryClientProvider } from '@tanstack/react-query'
 
@@ -8,7 +6,6 @@ import { routeTree } from './routeTree.gen'
 import './lib/i18n'
 
 import './styles.css'
-import reportWebVitals from './reportWebVitals.ts'
 
 import './interceptor-setup'
 
@@ -31,19 +28,10 @@ declare module '@tanstack/react-router' {
   }
 }
 
-// Render the app
-const rootElement = document.getElementById('app')!
-const root = ReactDOM.createRoot(rootElement)
-root.render(
-<StrictMode>
-  <QueryClientProvider client={queryClient}>
-    <RouterProvider router={router} />
-  </QueryClientProvider>
-</StrictMode>,
-)
-
-
-// If you want to start measuring performance in your app, pass a function
-// to log results (for example: reportWebVitals(console.log))
-// or send to an analytics endpoint. Learn more: https://bit.ly/CRA-vitals
-reportWebVitals()
+export function App() {
+  return (
+    <QueryClientProvider client={queryClient}>
+      <RouterProvider router={router} />
+    </QueryClientProvider>
+  )
+}
