+++ frontend/src/index.tsx
@@ -1,0 +1,18 @@
+import { StrictMode } from 'react'
+import ReactDOM from 'react-dom/client'
+import reportWebVitals from './reportWebVitals.ts'
+import { App } from './App.tsx'
+
+// Render the app
+const rootElement = document.getElementById('app')!
+const root = ReactDOM.createRoot(rootElement)
+root.render(
+  <StrictMode>
+    <App />
+  </StrictMode>,
+)
+
+// If you want to start measuring performance in your app, pass a function
+// to log results (for example: reportWebVitals(console.log))
+// or send to an analytics endpoint. Learn more: https://bit.ly/CRA-vitals
+reportWebVitals()
